+++ Back/src/main/java/iut/unilim/fr/back/dto/admin/MCCCResourceDTO.java
@@ -280,12 +280,12 @@ public class MCCCResourceDTO {
     public static class UECoefficientDTO {
         private String ueLabel;        // e.g., "UE2.1"
         private String ueName;         // e.g., "Understanding organizations"
-        private Integer coefficient;
+        private Double coefficient;
 
         public UECoefficientDTO() {
         }
 
-        public UECoefficientDTO(String ueLabel, String ueName, Integer coefficient) {
+        public UECoefficientDTO(String ueLabel, String ueName, Double coefficient) {
             this.ueLabel = ueLabel;
             this.ueName = ueName;
             this.coefficient = coefficient;
@@ -307,11 +307,11 @@ public class MCCCResourceDTO {
             this.ueName = ueName;
         }
 
-        public Integer getCoefficient() {
+        public Double getCoefficient() {
             return coefficient;
         }
 
-        public void setCoefficient(Integer coefficient) {
+        public void setCoefficient(Double coefficient) {
             this.coefficient = coefficient;
         }
     }
